:heavy_minus_sign: Remove meatie from the project and reimplement the S3I interfacing using httpx.
Meatie has proven to be more something to fiddle with than something ready for real world use. I will continue working with the meatie maintainer to try to make it a great project.

In the process of rewriting, I fixed various bugs related to 1. the auth flow not implementing the correct method, 2. various headaches with using TypeAdapters of TypeAdapters and so on, 3. much more (-:

I am so happy I finally got this somewhat working. This is not tested completely right now though!