🗃️ Fix Device link not being created properly in `new_image` and `status` handlers
This issue occured because the device link type was `Link[Device] | Device` in both cases to avoid issues with pyright if I recall correctly. This however led to Beanie not serializing the links anymore.

I solved this issue by removing the `| Device` union and ignoring the pyright errors via comments.

Closes #7