:loud_sound: Incorporate the callsite in the log message
I moved from looking up the callsite using stack inspection to just using the logger name. This should not only make things easier to read, but way more performant (if this would be a criteria at this point ...). For that, I also needed to construct my own column dictionary. Aside from that, I restructured the code a bit to make it more readable.