:bug: Move `forever` subcommand of `exp-coord run` into it's own click
This has to be done as now the callback on the level above is also executed when running `exp-coord run forever`. I achieved building a single-command click which behaves just like a `@app.command()` by putting the code into the callback of the `forever_app` and setting `invoke_without_command` to `True`. Not nice, but it works (-: